Add the support for Vulkan 1.3
This patch will be applied to frameworks/native/ .

Currently, the vulkaninfo will show the Vulkan instance version is 1.1.
Actually, our mesa implementation already supports version 1.3.
This change will help to fix it.
Note,the reason I'm not using VK_API_VERSION_1_3 here is that
it hasn't been defined in prebuilts/xxx/vulkan_core.h yet.
In the latest AOSP code, vulkan_core.h will include this definition,
this code will also be updated to use "*pApiVersion = VK_API_VERSION_1_3;"
in the commit of '628c41a'.
When we upgrade to the latest AOSP version, we need to remove this patch.

Tracked-On: OAM-115564
Signed-off-by: Li,Gaoshun gaoshun.li@intel.com